settings.py: Fix Django compatibility issue with TEMPLATE_CONTEXT_PROCESSORS
Instead of defining all the TEMPLATE_CONTEXT_PROCESSORS in settings.py, the
setting was changed to use Django's default context processors and add on
any additional context processors needed for this project. The alternative
leads to incompatibility issues when the default context processors are
renamed (as the auth context processor was, for Django 1.2). This fix is not
likely to cause any issues.

For more information:

https://docs.djangoproject.com/en/1.2/ref/settings/#template-context-processors
http://blog.madpython.com/2010/04/07/django-context-processors-best-practice/