Add line-number alignment for snippet display
Now, the line number lines up with the line itself, even if the line is
longer than the width of the box and even if you zoom in. Since this
change removes the need for the toggle-wordwrap feature, that has been
removed. Also, the line numbers are now right-aligned, and the pre tags
have been removed, allowing for neater indentation in the template file.
(The useful attributes of <pre> are now being applied via CSS.)

There is one minor cosmetic issue: the corners of the snippet display
box look a bit fuzzy when the background of the box is dark. Not quite
sure how to fix it at the moment.