Add versioning to bundles
Some other bundle-related changes were made, including:

* Editing the snippetform CSS and HTML to allow bundle/form.djhtml to
  be reused for editing
* Changing {% block title %} to {% block section %} in the base
  template for bundles to allow for more flexibility when creating
  breadcrumbs
* Saved common URL patterns in variables in bundle/urls.py
* Renamed explore.html to explore.djhtml for consistency

You should now be able to upload new versions as well as view the files
(or a particular file) for a bundle at a specific version.

Coming soon: the ability to add a timestamp and a comment for each new
uploaded version (if this feature is desirable).